@@ -1,5 +1,3 @@
-import 'dart:io';
-
 import 'package:audioplayers/audioplayers.dart';
 import 'package:flutter/foundation.dart';
 import 'package:flutter_test/flutter_test.dart';
@@ -10,12 +8,11 @@ import 'lib/lib_test_utils.dart';
 import 'platform_features.dart';
 import 'test_utils.dart';
 
-void main() {
-  final features = PlatformFeatures.instance();
-
+void main() async {
   IntegrationTestWidgetsFlutterBinding.ensureInitialized();
-
-  final isAndroid = !kIsWeb && Platform.isAndroid;
+  final features = PlatformFeatures.instance();
+  final isAndroid = !kIsWeb && defaultTargetPlatform == TargetPlatform.android;
+  final audioTestDataList = await getAudioTestDataList();
 
   group('play multiple sources', () {
     testWidgets(
@@ -35,7 +32,7 @@ void main() {
         await tester.pump(const Duration(seconds: 8));
         for (var i = 0; i < audioTestDataList.length; i++) {
           final td = audioTestDataList[i];
-          if (td.isLiveStream || td.duration > const Duration(seconds: 10)) {
+          if (td.isLiveStream || td.duration! > const Duration(seconds: 10)) {
             await tester.pump();
             final position = await players[i].getCurrentPosition();
             printWithTimeOnFailure('Test position: $td');
@@ -56,14 +53,13 @@ void main() {
         (WidgetTester tester) async {
       final player = AudioPlayer();
 
-      for (var i = 0; i < audioTestDataList.length; i++) {
-        final td = audioTestDataList[i];
+      for (final td in audioTestDataList) {
         await tester.pumpLinux();
         await player.play(td.source);
         await tester.pumpAndSettle();
         // Sources take some time to get initialized
         await tester.pump(const Duration(seconds: 8));
-        if (td.isLiveStream || td.duration > const Duration(seconds: 10)) {
+        if (td.isLiveStream || td.duration! > const Duration(seconds: 10)) {
           await tester.pump();
           final position = await player.getCurrentPosition();
           printWithTimeOnFailure('Test position: $td');
@@ -101,7 +97,8 @@ void main() {
         await tester.pumpLinux();
         await player.play(td.source);
         await tester.pumpAndSettle();
-        await tester.pump(td.duration + const Duration(seconds: 8));
+        await tester
+            .pump((td.duration ?? Duration.zero) + const Duration(seconds: 8));
         expect(player.state, PlayerState.completed);
 
         audioContext = AudioContextConfig(
@@ -114,7 +111,8 @@ void main() {
 
         await player.resume();
         await tester.pumpAndSettle();
-        await tester.pump(td.duration + const Duration(seconds: 8));
+        await tester
+            .pump((td.duration ?? Duration.zero) + const Duration(seconds: 8));
         expect(player.state, PlayerState.completed);
         await tester.pumpLinux();
         await player.dispose();
@@ -148,7 +146,8 @@ void main() {
         await player.setSource(td.source);
         await player.resume();
         await tester.pumpAndSettle();
-        await tester.pump(td.duration + const Duration(seconds: 8));
+        await tester
+            .pump((td.duration ?? Duration.zero) + const Duration(seconds: 8));
         expect(player.state, PlayerState.playing);
         await player.stop();
         expect(player.state, PlayerState.stopped);
@@ -163,7 +162,8 @@ void main() {
 
         await player.resume();
         await tester.pumpAndSettle();
-        await tester.pump(td.duration + const Duration(seconds: 8));
+        await tester
+            .pump((td.duration ?? Duration.zero) + const Duration(seconds: 8));
         expect(player.state, PlayerState.playing);
         await player.stop();
         expect(player.state, PlayerState.stopped);
