@@ -39,19 +39,10 @@ class Dlg extends StatelessWidget {
   @override
   Widget build(BuildContext context) {
     return Dialog(
-      shape: RoundedRectangleBorder(
-        borderRadius: BorderRadius.circular(8.0),
-      ),
-      elevation: 0,
-      backgroundColor: Colors.white,
-      child: Container(
+      child: Padding(
         padding: const EdgeInsets.all(16.0),
-        child: contentBox(context),
+        child: child,
       ),
     );
-  }
-
-  Widget contentBox(BuildContext context) {
-    return child;
   }
 }
