@@ -2,12 +2,16 @@ package xyz.luan.audioplayers
 
 import android.annotation.SuppressLint
 import android.media.AudioAttributes
-import android.media.AudioAttributes.*
+import android.media.AudioAttributes.Builder
+import android.media.AudioAttributes.CONTENT_TYPE_MUSIC
+import android.media.AudioAttributes.USAGE_MEDIA
+import android.media.AudioAttributes.USAGE_NOTIFICATION_RINGTONE
+import android.media.AudioAttributes.USAGE_VOICE_COMMUNICATION
 import android.media.AudioManager
 import android.media.MediaPlayer
 import android.os.Build
 import androidx.annotation.RequiresApi
-import java.util.Objects
+import java.util.*
 
 data class AudioContextAndroid(
     val isSpeakerphoneOn: Boolean,
@@ -55,11 +59,11 @@ data class AudioContextAndroid(
 
     override fun hashCode() = Objects.hash(isSpeakerphoneOn, stayAwake, contentType, usageType, audioFocus, audioMode)
 
-    override fun equals(other: Any?) = (other is AudioContextAndroid)
-            && isSpeakerphoneOn == other.isSpeakerphoneOn
-            && stayAwake == other.stayAwake
-            && contentType == other.contentType
-            && usageType == other.usageType
-            && audioFocus == other.audioFocus
-            && audioMode == other.audioMode
+    override fun equals(other: Any?) = (other is AudioContextAndroid) &&
+        isSpeakerphoneOn == other.isSpeakerphoneOn &&
+        stayAwake == other.stayAwake &&
+        contentType == other.contentType &&
+        usageType == other.usageType &&
+        audioFocus == other.audioFocus &&
+        audioMode == other.audioMode
 }
