@@ -1,6 +1,5 @@
 package xyz.luan.audioplayers
 
-
 import android.content.Context
 import android.media.AudioManager
 import android.os.Build
@@ -159,12 +158,12 @@ class AudioplayersPlugin : FlutterPlugin, IUpdateCallback {
                 }
 
                 "getDuration" -> {
-                    response.success(player.getDuration() ?: 0)
+                    response.success(player.getDuration())
                     return
                 }
 
                 "getCurrentPosition" -> {
-                    response.success(player.getCurrentPosition() ?: 0)
+                    response.success(player.getCurrentPosition())
                     return
                 }
 
@@ -232,7 +231,7 @@ class AudioplayersPlugin : FlutterPlugin, IUpdateCallback {
         handler.post {
             player.eventHandler.success(
                 "audio.onDuration",
-                hashMapOf("value" to (player.getDuration() ?: 0))
+                hashMapOf("value" to (player.getDuration() ?: 0)),
             )
         }
     }
@@ -265,7 +264,8 @@ class AudioplayersPlugin : FlutterPlugin, IUpdateCallback {
         handler.post {
             player.eventHandler.success("audio.onSeekComplete")
             player.eventHandler.success(
-                "audio.onCurrentPosition", hashMapOf("value" to (player.getCurrentPosition() ?: 0))
+                "audio.onCurrentPosition",
+                hashMapOf("value" to (player.getCurrentPosition() ?: 0)),
             )
         }
     }
