@@ -107,10 +107,11 @@ class AudioContextConfig {
     );
   }
 
-  AudioContextIOS buildIOS() {
-    if (defaultTargetPlatform == TargetPlatform.iOS) {
-      validateIOS();
+  AudioContextIOS? buildIOS() {
+    if (defaultTargetPlatform != TargetPlatform.iOS) {
+      return null;
     }
+    validateIOS();
     return AudioContextIOS(
       category: respectSilence
           ? AVAudioSessionCategory.ambient
@@ -119,22 +120,37 @@ class AudioContextConfig {
               : (route == AudioContextConfigRoute.earpiece
                   ? AVAudioSessionCategory.playAndRecord
                   : AVAudioSessionCategory.playback)),
-      options: (duckAudio
-              ? [AVAudioSessionOptions.duckOthers]
-              : <AVAudioSessionOptions>[]) +
-          (route == AudioContextConfigRoute.speaker
-              ? [AVAudioSessionOptions.defaultToSpeaker]
-              : []),
+      options: {
+        if (duckAudio) AVAudioSessionOptions.duckOthers,
+        if (route == AudioContextConfigRoute.speaker)
+          AVAudioSessionOptions.defaultToSpeaker,
+      },
     );
   }
 
   void validateIOS() {
-    // Please create a custom [AudioContextIOS] if the generic flags cannot
-    // represent your needs.
-    if (respectSilence && route == AudioContextConfigRoute.speaker) {
-      throw 'On iOS it is impossible to set both respectSilence and '
-          'forceSpeaker';
-    }
+    const invalidMsg =
+        'Invalid AudioContextConfig: On iOS it is not possible to set';
+    const tip = 'Please create a custom [AudioContextIOS] if the generic flags '
+        'cannot represent your needs.';
+    assert(
+      !(duckAudio && respectSilence),
+      '$invalidMsg `respectSilence` and `duckAudio`. $tip',
+    );
+    assert(
+      !(respectSilence && route == AudioContextConfigRoute.speaker),
+      '$invalidMsg `respectSilence` and route `speaker`. $tip',
+    );
+  }
+
+  @override
+  String toString() {
+    return 'AudioContextConfig('
+        'route: $route, '
+        'duckAudio: $duckAudio, '
+        'respectSilence: $respectSilence, '
+        'stayAwake: $stayAwake'
+        ')';
   }
 }
 
